review comments
* reorder market field inits

* relocate test to bing
